refactor(output): extract result_types tests to sibling file
